switched nextjs back to app router layout and started github grid component
