+++ app/api/notebook/deleteNote/noteId/[noteId].ts
