+++ app/components/MainComponents/Hero.tsx
@@ -27,4 +27,4 @@ const Hero = () => {
   );
 }
 
-export default Hero
+export default Hero;
