+++ app/components/VetsWhoCode_Components/StarWarsComponents/CharacterData.tsx
