+++ app/page.tsx
@@ -1,0 +1,11 @@
+import { Metadata } from "next";
+import Main from "./components/MainComponents/Main";
+
+export const metadata: Metadata = {
+  description: "Jason Vallery's Portfolio",
+  title: "Jason Vallery's Portfolio",
+};
+
+export default function Home() {
+  return <Main />;
+}
