--- pages/_app.tsx
@@ -1,11 +1,0 @@
-import "@/styles/globals.css";
-import type { AppProps } from "next/app";
-import Layout from "../components/layout"
-
-export default function App({ Component, pageProps }: AppProps) {
-  return (
-    <Layout>
-      <Component {...pageProps} />
-    </Layout>
-  );
-}
